@@ -2,15 +2,16 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <!-- <EnableNETAnalyzers>true</EnableNETAnalyzers> -->
-    
+    <EnableNETAnalyzers>true</EnableNETAnalyzers>
+    <!-- <AnalysisLevel>latest</AnalysisLevel>
+    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild> -->
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Dumpify" Version="0.6.6" />
   </ItemGroup>
-
+    
 </Project>
